server/pyproject.toml
[tool.poetry]
name = "eleden-server"
version = "1.8.0"
description = "electronic education environment"
authors = ["Luferov Victor <lyferov@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.10.0"
celery = "^5.2.3"
channels = "^3.0.4"
channels-redis = "^3.3.1"
daphne = "^3.0.2"
Django = "^3"
django-celery-beat = "^2.2.1"
django-celery-results = "^2.2.0"
django-channels-graphql-ws = "^0.9.1"
django-cors-headers = "^3.11.0"
django-cors-middleware = "^1.5.0"
django-filter = "^21.1"
django-push-notifications = "^3.0.0"
djangorestframework = "^3.13.1"
flatten-dict = "^0.4.2"
graphene-django = "^2.15.0"
graphene-django-optimizer = "^0.8"
graphene-file-upload = "^1.3.0"
graphql-relay = "^2"
openpyxl = "^3.0.9"
Pillow = "^9.0.1"
progress = "^1.6"
python-dotenv = "^0.19.2"
pywebpush = "^1.14.0"
redis = "^4.1.2"
requests = "^2.27.1"
sentry-sdk = "^1.5.4"
user-agents = "^2.2.0"
bibtexparser = "^1.2.0"
psycopg2-binary = "^2.9.2"
django-oauth-toolkit = "^1.7.0"
paramiko = "^2.8.1"
lxml = "4.8.0"
inflection = "^0.5.1"
devind-helpers = "^0.2"
devind-core = "^0.1.3"
devind-notifications = "^0.1.0"
django-cbias-auth = "^0.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
flake8-docstrings = "^1.6.0"
flake8-builtins = "^1.5.3"
flake8-quotes = "^3.3.1"
flake8-comprehensions = "^3.8.0"
flake8-eradicate = "^1.2.0"
flake8-simplify = "^0.18.1"
flake8-spellcheck = "^0.25.0"
pep8-naming = "^0.12.1"
flake8-use-fstring = "^1.3"
flake8-annotations = "^2.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
